multimodal-lambda-architecture-master/requirements.txt
pyspark==3.5.1
kafka-python==2.0.2
cassandra-driver==3.29.0
Flask==3.0.2
pandas==2.2.3
pyarrow==18.0.0
